Updated GameManager.java
Changed the currentPlayer object so it contains a int instead of a player object, also renamed it to currentTurn. Updated the switchTurns method to reflect the change.

Added a method called checkForGameOver. It runs through a player's fleet and checks if any of them have not been sunk yet. Returns false if at least one ship has not been sunk yet, otherwise it returns true.